asus-switcheroo: i915 jprobe hack
This is an attempt to work around i915 bugs when we switch over
to discrete graphics.  It's far from perfect, needing to be loaded
before i915 in the initramfs and hitting a BUG while registering
a jprobe, but it seems to do the job.  Not even attempting to make
this part of the install yet.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>