asus-switcheroo: Make nouveau reprobe
Only newer kernels support the reprobe vga switcheroo callback, but
this is the thing we need to get output going when we switch to
discrete graphics.  The function we need isn't exported, so go
through a sneaky kallsym trick to call it anyway.  With this,
we get output on the LCD when we switch to nouveau discrete gfx.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>