asus-switcheroo: more fun with jprobes (and kretprobes!)
When only using asus-switcheroo to power off the discrete graphics,
it's not uncommon for the system to eventually lock up.  On my
system, this often seems to be a result of nouveau getting an
interrupt (it's on a shared irq, so it should expect this).  The
interrupt handler goes out to read hardware, gets -1 back and
proceeds to kill the system.  These probes free the interrupt
handler prior to putting the device in D3hot, and re-request it
after the device is returned to D0.  I've already submitted a
patch upstream to do exactly this, but it hasn't yet been well
received.  Like the i915-jprobe, this needs to be loaded prior
to the driver, nouveau in this case.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>